websites/AttendanceLog/requirements.txt
Faker==19.3.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-Minify==0.42
twilio==8.9.1
psycopg2-binary==2.9.7
werkzeug==2.3.6
python-dotenv==1.0.0
pylightxl==1.61